Migrate fjs/djs/tokenizer to .f.mjs
Implementation and proof migrate from authored TypeScript to
JSDoc-typed JavaScript under the stage-1 migration. DjsToken and
DjsTokenWithMetadata move to a sibling types.ts since fjs/djs/parser
depends on them. Private types (Token, FlatToken, TokenScanState,
StringDecodeState, DjsScanState) stay as _-prefixed JSDoc typedefs in
the module. generic descentParser<TokenMetadata>(...) calls, which JS
has no syntax for, get an inline cast on the result instead
(DescentMatch<TokenMetadata>). Updates fjs/djs/parser/module.f.ts,
fjs/djs/parser/proof.f.ts, and fjs/djs/transpiler/module.f.ts import
specifiers, and fixes stale module.f.ts/proof.f.ts doc references
across six todo docs — several already had stale line-number citations
predating this migration (verified against the pre-migration file), so
only the .f.ts -> .f.mjs extension was corrected there; references
whose line numbers still matched (or could be confidently recomputed)
were fixed exactly.
