+++ fjs/djs/parser/module.f.ts
@@ -8,7 +8,7 @@ import { error, ok } from '../../types/result/module.f.mjs'
 import type { List } from '../../types/list/types.ts'
 import { fold, next, toArray, length, concat } from '../../types/list/module.f.mjs'
 import type { Fold } from '../../types/function/operator/types.ts'
-import type { DjsToken, DjsTokenWithMetadata } from '../tokenizer/module.f.ts'
+import type { DjsToken, DjsTokenWithMetadata } from '../tokenizer/types.ts'
 import { setReplace, at } from '../../types/ordered_map/module.f.mjs'
 import type { OrderedMap } from '../../types/ordered_map/types.ts'
 import { fromMap } from '../../types/object/module.f.mjs'
