+++ fjs/djs/transpiler/module.f.ts
@@ -8,7 +8,7 @@ import type { Result } from '../../types/result/types.ts'
 import { error, ok } from '../../types/result/module.f.mjs'
 import type { List } from '../../types/list/types.ts'
 import { drop, map as listMap, toArray, includes } from '../../types/list/module.f.mjs'
-import { tokenize } from '../tokenizer/module.f.ts'
+import { tokenize } from '../tokenizer/module.f.mjs'
 import { setReplace, at } from '../../types/ordered_map/module.f.mjs'
 import type { OrderedMap } from '../../types/ordered_map/types.ts'
 import { stringToList } from '../../text/utf16/module.f.mjs'
